% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/container.R
\name{as.container}
\alias{as.container}
\title{Coerce an object into a container object.}
\usage{
as.container(x, host = localhost)
}
\description{
A container object represents a Docker container on a host.
}

