% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{as.image}
\alias{as.image}
\alias{as.image.character}
\alias{as.image.list}
\alias{as.image.image}
\title{Coerce an object into an image object.}
\usage{
as.image(x, host = harbor::localhost)

\method{as.image}{character}(x, host = harbor::localhost)

\method{as.image}{list}(x, host = harbor::localhost)

\method{as.image}{image}(x, host = harbor::localhost)
}
\arguments{
\item{x}{object to convert}

\item{host}{docker host}
}
\description{
Coerce an object into an image object.
}
