% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2docker.R
\name{boot2docker_shellinit}
\alias{boot2docker_shellinit}
\title{If we're on Mac and Windows, we're using boot2docker, and we need to run the
equivalent of `$(boot2docker shellinit)`.}
\usage{
boot2docker_shellinit()
}
\description{
If we're on Mac and Windows, we're using boot2docker, and we need to run the
equivalent of `$(boot2docker shellinit)`.
}
\keyword{internal}

