% Generated by roxygen2 (4.0.2): do not edit by hand
\name{container_rm}
\alias{container_rm}
\title{Delete a container.}
\usage{
container_rm(container, force = FALSE)
}
\arguments{
\item{force}{Force removal of a running container.}
}
\description{
Delete a container.
}
\examples{
\dontrun{
container_rm(con)
}
}

