% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/container.R
\name{container_running}
\alias{container_running}
\title{Report whether a container is currently running.}
\usage{
container_running(container)
}
\description{
Report whether a container is currently running.
}
\examples{
\dontrun{
container_running(con)
}
}

