% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{container_stop}
\alias{container_stop}
\title{Stop a running container}
\usage{
container_stop(container, seconds = 10)
}
\arguments{
\item{container}{A container object}

\item{seconds}{Seconds to wait for stop before killing container}
}
\description{
Stop a running container
}
\examples{
\dontrun{
container_stop(con)
}
}
