% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet.R
\name{docker_cmd.droplet}
\alias{docker_cmd.droplet}
\title{Integrate with Docker endpoints on Digital Ocean}
\usage{
\method{docker_cmd}{droplet}(host, cmd = NULL, args = NULL,
  docker_opts = NULL, capture_text = FALSE, ...)
}
\arguments{
\item{host}{reachable hostname}

\item{cmd}{docker command}

\item{args}{args to docker command}

\item{docker_opts}{options passed to tocker}

\item{capture_text}{should we capture output text? default: \code{FALSE}}

\item{...}{additional options passed to \link[analogsea:droplet_ssh]{analogsea::droplet_ssh} and
\link[analogsea:droplet_download]{analogsea::droplet_download}s}
}
\description{
Integrate with Docker endpoints on Digital Ocean
}
