% Generated by roxygen2 (4.0.2): do not edit by hand
\name{docker_cmd}
\alias{docker_cmd}
\title{Run a docker command on a host.}
\usage{
docker_cmd(host, cmd = NULL, args = NULL, docker_opts = NULL, ...)
}
\description{
Run a docker command on a host.
}
\examples{
\dontrun{
docker_cmd(localhost(), "ps", "-a")
}
}

