% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docker-machine.R
\name{docker_machine_init}
\alias{docker_machine_init}
\title{Connect to a docker machine}
\usage{
docker_machine_init(machine = NULL)
}
\arguments{
\item{machine}{Name of a machine}
}
\description{
Initialise environment variables to connect to a docker machine.
}
\details{
In the case where there is only a single docker machine and no
machine is specified there is no difficult logic.

\enumerate{
\item{If \code{machine} is not specified (or is \code{NULL}) then
\enumerate{
\item{If \code{DOCKER_MACHINE_NAME} is set (i.e., \emph{something}
has already set \code{docker-machine} variables) do nothing and
use these.}

\item{Otherwise pick the first name listed among running machines
(see \code{docker ls --filter state=Running}) with a message if
more than one was found}
}}

\item{If \code{machine} is specified and non-\code{NULL} then use
this machine only, throwing an error if it is not running}
}
}

