% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_tag}
\alias{docker_tag}
\title{Create a target image that refers to a source image}
\usage{
docker_tag(host = harbor::localhost, source_image = NULL,
  target_image = NULL, capture_text = FALSE)
}
\arguments{
\item{host}{docker host}

\item{source_image}{source image object or name}

\item{target_image}{target image name}

\item{capture_text}{capture text of output?}
}
\description{
Create a target image that refers to a source image
}
