% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbor-package.R
\docType{package}
\name{harbor}
\alias{harbor}
\alias{harbor-package}
\title{Tools to Manage 'docker' Containers}
\description{
Tools to Manage 'docker' Containers
}
\examples{
\dontrun{
docker_pull(image="hello-world")
images() # find the one we just downloaded & remember the name/id
docker_run(image="imgname")
}
}
