% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{image_rm}
\alias{image_rm}
\title{Delete an image}
\usage{
image_rm(x, force = FALSE)
}
\arguments{
\item{x}{An image object}

\item{force}{Force removal of a running container.}
}
\description{
Delete an image
}
\examples{
\dontrun{
image_rm(img)
}
}
