% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{is_container}
\alias{is_container}
\title{Test to see if an object is a docker container object}
\usage{
is_container(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical
}
\description{
Test to see if an object is a docker container object
}
