% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{is_image}
\alias{is_image}
\title{Test to see if an object is a docker image object}
\usage{
is_image(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical
}
\description{
Test to see if an object is a docker image object
}
