% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docker-machine.R
\name{localhost_ip}
\alias{localhost_ip}
\title{IP address of docker server}
\usage{
localhost_ip(machine = NULL)
}
\arguments{
\item{machine}{Name of the machine}
}
\description{
Return the IP address for the "local" machine.  On Linux this is
just the localhost, but on Mac and Windows system it will be the
IP of the docker-machine VM.
}

