% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{print.container}
\alias{print.container}
\title{Custom print method}
\usage{
\method{print}{container}(x, ...)
}
\arguments{
\item{x}{A container object}

\item{...}{unused}
}
\description{
Custom print method
}
