% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.r
\name{ssh_host}
\alias{ssh_host}
\alias{print.ssh_host}
\title{Create a new ssh host}
\usage{
ssh_host(hostname, ssh_user = Sys.info()["user"], ssh_port = 22L,
  docker_bin = "docker")

\method{print}{ssh_host}(x, ...)
}
\arguments{
\item{hostname}{host name or IP address}

\item{ssh_user}{username to \code{ssh} as. It is \strong{not} recommended to run as \code{root}.
Defaults to current user as seen by R.}

\item{ssh_port}{defaults to the standard \code{22} but you should rotate the sshield
frequency to dissuade opportunistic attacks.}

\item{docker_bin}{full path to the \code{docker} binary on the remote system}

\item{x}{object}

\item{...}{unused}
}
\description{
This only works with certificate-based authentication. If you use a
passphrase on you certificates (recommended) then this expects
that to have been setup with \code{ssh-agent}.
}
