% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.r
\name{version}
\alias{version}
\alias{print.dckr_version}
\title{Get Docker version info}
\usage{
version(host = harbor::localhost)

\method{print}{dckr_version}(x, ...)
}
\arguments{
\item{host}{docker host}

\item{x}{object}

\item{...}{unused}
}
\description{
Get Docker version info
}
