% Generated by roxygen2 (4.0.2): do not edit by hand
\name{container_logs}
\alias{container_logs}
\title{Retrieve logs for a container.}
\usage{
container_logs(container, timestamps = FALSE, follow = FALSE)
}
\arguments{
\item{follow}{Follow log output as it is happening.}

\item{timestamp}{Show timestamps.}
}
\description{
Retrieve logs for a container.
}
\examples{
\dontrun{
container_rm(con)
}
}

