% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{container_logs}
\alias{container_logs}
\title{Retrieve logs for a container.}
\usage{
container_logs(container, timestamps = FALSE)
}
\arguments{
\item{container}{A container object}

\item{timestamps}{Show timestamps.}
}
\description{
Retrieve logs for a container.
}
\examples{
\dontrun{
container_logs(con)
}
}
