% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R, R/localhost.R, R/ssh.r
\name{docker_cmd}
\alias{docker_cmd}
\alias{docker_cmd.localhost}
\alias{docker_cmd.ssh_host}
\title{Run a docker command on a host.}
\usage{
docker_cmd(host = harbor::localhost, cmd = NULL, args = NULL,
  docker_opts = NULL, capture_text = FALSE, ...)

\method{docker_cmd}{localhost}(host, cmd = NULL, args = NULL,
  docker_opts = NULL, capture_text = FALSE, text_from = "stdout", ...)

\method{docker_cmd}{ssh_host}(host, cmd = NULL, args = NULL,
  docker_opts = NULL, capture_text = FALSE, text_from = "stdout", ...)
}
\arguments{
\item{host}{A host object.}

\item{cmd}{A docker command, such as "run" or "ps"}

\item{args}{Arguments to pass to the docker command. These are passed in
the same way they would need to be passed in to \code{sys::exec_} calls.
i.e. "\code{-q -a -s}" would be \code{c("-q", "-a", "-s")}.}

\item{docker_opts}{Options to docker. These are things that come before the
docker command, when run on the command line.}

\item{capture_text}{If \code{FALSE} (the default), return the host object.
This is useful for chaining functions. If \code{TRUE``, capture the text output from both stdout and stderr, and return that. Note that}TRUE``
may not be available on all types of hosts.}

\item{...}{Other arguments passed to the SSH command for the host}
}
\description{
Run a docker command on a host.
}
\examples{
\dontrun{
docker_cmd(localhost, "ps", "-a")
}
}
