% Generated by roxygen2 (4.0.2): do not edit by hand
\name{docker_pull}
\alias{docker_pull}
\title{Pull a docker image onto a host.}
\usage{
docker_pull(host = localhost, image, ...)
}
\value{
The \code{host} object.
}
\description{
Pull a docker image onto a host.
}
\examples{
\dontrun{
docker_pull(localhost, "debian:testing")
}
}

