% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{localhost}
\alias{localhost}
\title{An object representing the current computer that R is running on.}
\format{\preformatted{ list()
 - attr(*, "class")= chr [1:2] "localhost" "host"
}}
\usage{
localhost
}
\description{
An object representing the current computer that R is running on.
}
\keyword{datasets}

