% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.r
\name{version}
\alias{version}
\alias{print.dversion}
\title{Get Docker version info}
\usage{
version(host = harbor::localhost)

\method{print}{dversion}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{unused}

\item{docker}{host}
}
\description{
Get Docker version info
}
